PMM-6793 add FormParts styles
