+++ pmm-app/src/pmm-add-instance-app-panel/components/AddRemoteInstance/FormParts/FormParts.styles.tsx
@@ -1,0 +1,9 @@
+import { stylesFactory } from '@grafana/ui';
+import { GrafanaTheme } from '@grafana/data';
+import { css } from 'emotion';
+
+export const getStyles = stylesFactory((theme: GrafanaTheme) => ({
+  groupWrapper: css`
+    width: 50%;
+  `,
+}));
